Minor fixes and create file for MaxCensusSizeSelector component
